requirements.txt
fastapi
uvicorn
requests
google-generativeai
speechrecognition
pyttsx3
pyaudio
selenium
webdriver-manager
pyautogui
python-dotenv
pywebview
